pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logteehtml"
dynamic = ["version"]
description = "Comprehensive HTML logging with content support"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "André Aichert", email = "aaichert@gmail.com"}
]
maintainers = [
    {name = "André Aichert", email = "aaichert@gmail.com"}
]
keywords = ["logging", "html", "pillow", "rich", "documentation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Logging",
    "Topic :: Text Processing :: Markup :: HTML",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "Pillow>=8.0.0",
    "rich>=10.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=6.0.0", "black", "flake8", "build", "twine"]

[project.urls]
Homepage = "https://github.com/aaichert/logteehtml"
Documentation = "https://github.com/aaichert/logteehtml#readme"
Repository = "https://github.com/aaichert/logteehtml.git"
"Bug Tracker" = "https://github.com/aaichert/logteehtml/issues"

[tool.setuptools.dynamic]
version = {attr = "logteehtml.__version__.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["logteehtml*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
